Refactor Manifest and Package Structure
