+++ app/src/main/AndroidManifest.xml
@@ -11,22 +11,12 @@
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
-        android:theme="@style/Theme.DateMee"
+        android:theme="@style/Theme.DateMe"
         tools:targetApi="31">
         <activity
-            android:name=".Onboarding"
-            android:exported="false"
-            android:label="@string/title_activity_onboarding"
-            android:theme="@style/Theme.DateMee">
-            <meta-data
-                android:name="android.app.lib_name"
-                android:value="" />
-        </activity>
-        <activity
-            android:name=".MainActivity"
+            android:name="com.androiders.dateme.features.onboading.presentation.MainActivity"
             android:exported="true"
-            android:label="@string/app_name"
-            android:theme="@style/Theme.DateMee">
+            android:theme="@style/Theme.DateMe">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
