--- app/src/main/java/com/androiders/dateme/CopyContainer.kt
@@ -1,30 +1,0 @@
-package com.androiders.dateme
-
-/*
-            Text(
-                buildAnnotatedString {
-                    withStyle(style = ParagraphStyle(lineHeight = 30.sp)) {
-                        withStyle(
-                            style = SpanStyle(
-                                fontWeight = FontWeight.Bold,
-                                fontSize = 55.sp,
-                                color = MaterialTheme.colors.primary)
-
-                        ) {
-                            append("Hello\n")
-                        }
-                        withStyle(
-                            style = SpanStyle(
-                                fontWeight = FontWeight.Bold,
-                                fontSize = 25.sp,
-                                color = MaterialTheme.colors.primary
-                            )
-                        ) {
-                            append("DatingMee\n")
-                        }
-                        append("App made")
-                    }
-                }
-            )
-
- */
