+++ app/src/main/java/com/androiders/dateme/core/theme/Color.kt
@@ -1,4 +1,4 @@
-package com.androiders.dateme.core.ui.theme
+package com.androiders.dateme.core.theme
 
 import androidx.compose.ui.graphics.Color
